v0.1.1
Game is really simple for now. pressing space with make the "hero" (cube for now) to move-jump from car to car. pressing z' will make the hero to jump to the other faster/slower lanes. If its game over you can press 'R to restart. there is no menu or explainations to the game in this version. this version is the first 'playable' version with purpose(cahllenge yourself to do a high score).